+++ db/migrate/20140129195252_add_degree_or_certifications_earned_to_annuals.rb
@@ -1,0 +1,5 @@
+class AddDegreeOrCertificationsEarnedToAnnuals < ActiveRecord::Migration
+  def change
+    add_column :annuals, :degree_or_certifications_earned, :text
+  end
+end